refactor: Organize MCP components into correct architectural layers
I've moved MCP-related domain components from the previous `ai_interaction_service`
structure into more specific modules under `novade-domain/src/ai/` and
`novade-domain/src/ai/mcp/`.

This includes:
- Relocation of types, client_instance, transport, connection_service,
  consent_manager, and logic_service files.
- Updates to `mod.rs` files in `novade-domain/src/ai/`,
  `novade-domain/src/ai/mcp/`, and `novade-domain/src/lib.rs` to reflect
  the new module hierarchy and ensure correct re-exports.
- Adjustments to internal `use` statements within the moved files to
  align with the new paths.

NOTE: I'm submitting this with a known unresolved external dependency
issue. The `novade-system` crate requires `smithay = "^0.10.0"`, which is
currently not resolvable in your build environment. This prevents
`novade-system` and consequently `novade-domain` from compiling. The
refactoring of module paths and structure within `novade-domain` is
structurally complete, but full compilation and testing are blocked pending
resolution of the `smithay` dependency.
